+++ phpunit.xml
@@ -15,4 +15,9 @@
             <directory>tests</directory>
         </testsuite>
     </testsuites>
+    <filter>
+    <whitelist processUncoveredFilesFromWhitelist="true">
+        <directory suffix=".php">src</directory>
+    </whitelist>
+    </filter>
 </phpunit>
