+++ tests/Collection/FindAllTest.php
@@ -1,0 +1,30 @@
+<?php
+
+use Collections\Collection;
+
+class FindAllTest extends PHPUnit_Framework_TestCase
+{
+    public function testFindAll()
+    {
+        $this->c = new Collection('TestClassA');
+        $this->c = $this->c->add(new TestClassA(54));
+        $this->c = $this->c->add(new TestClassA(32));
+        $this->c = $this->c->add(new TestClassA(32));
+        $this->c = $this->c->add(new TestClassA(32));
+
+        $condition = function ($item) {
+            return $item->getValue() == 32;
+        };
+
+        $subset = $this->c->filter($condition);
+
+        $c = (new Collection("TestClassA"))
+            ->add(new TestClassA(32))
+            ->add(new TestClassA(32))
+            ->add(new TestClassA(32));
+
+        $this->assertEquals(3, $subset->count());
+        $this->assertEquals($c, $subset);
+    }
+
+}
